Web Inspector: Really drop all locks in nested run loop on iOS if WebThread is enabled https://bugs.webkit.org/show_bug.cgi?id=130912
Patch by Joseph Pecoraro <pecoraro@apple.com> on 2014-03-28
Reviewed by Geoffrey Garen.

Previously we were calling DropAllLocks inside of a single line if statement,
so the JSLock was getting reaquired very quickly. We really want to DropAllLocks
for the duration of running the nested run loop on iOS if there is a WebThread.

* bindings/js/PageScriptDebugServer.h:
* bindings/js/PageScriptDebugServer.cpp:
(WebCore::PageScriptDebugServer::runEventLoopWhilePaused):
(WebCore::PageScriptDebugServer::runEventLoopWhilePausedInternal):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@166436 268f45cc-cd09-0410-ab3c-d52691b4dbfc