Don't prevent AppSleep from sync messages. https://bugs.webkit.org/show_bug.cgi?id=130747.
Reviewed by Anders Carlsson.

We think sync messages only need to happen full throttle when the plugin is visible,
the user is interacting with the plugin, or audio is playing.  Since we already
take assertions for all these cases the assertion in sync messages doesn’t provide
any value.  Since it can keep the PluginProcess awake in other instances remove it.

* PluginProcess/WebProcessConnection.cpp:
(WebKit::WebProcessConnection::didReceiveSyncMessage):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@166444 268f45cc-cd09-0410-ab3c-d52691b4dbfc