[Win] Update obsolete packages in cygwin https://bugs.webkit.org/show_bug.cgi?id=126264
Reviewed by Brent Fulgham.

Removed "apache" and "minires" as the first has been replaced by
"lighttpd" and the second is now integrated into
"cygwin". Replaced the transitional and now obsolete "gcc" package
with "gcc-g++" and ensure that "python" stays at version "2.6.8"
since the newer version available is "2.7.3" which breaks the
build due to some funniness with environment variable handling.

* CygwinDownloader/cygwin-downloader.py:
* CygwinDownloader/cygwin-downloader.zip:

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@166776 268f45cc-cd09-0410-ab3c-d52691b4dbfc