[CSSRegions] Fixed positioned elements in named flows have fragmented content https://bugs.webkit.org/show_bug.cgi?id=130595
Reviewed by Darin Adler.

Source/WebCore:

Fixed positioned elements that have the named flow as parent are positioned
and sized relative to the viewport. Because of that, when the named flow
has an associated region chain, the fixed positioned elements should not
be fragmented.

By making a fixed positioned element with a named flow parent unsplittable
for fragmentation, we prevent this wrong behavior.

Test: fast/regions/fixed-pos-content-fragmented.html

* rendering/RenderBox.cpp:
(WebCore::RenderBox::isUnsplittableForPagination):

LayoutTests:

* fast/regions/fixed-pos-content-fragmented-expected.html: Added.
* fast/regions/fixed-pos-content-fragmented.html: Added.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@166918 268f45cc-cd09-0410-ab3c-d52691b4dbfc