Make Vector::takeLast work with move-only types (and optimize for types where move is faster) https://bugs.webkit.org/show_bug.cgi?id=131735
Reviewed by Alexey Proskuryakov.

Source/WTF:
* wtf/Vector.h:
(WTF::Vector::takeLast): Added a missing std::move. All the other take functions have it.

Tools:
* TestWebKitAPI/Tests/WTF/Vector.cpp: Added a test for Vector::takeLast with move only types.
Fixed invocations of EXPECT macros so the expected value is on the left, and the test result
is on the right. This can be confusing since our JavaScript test framework does it the opposite
way, but gtest works better this way.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@167348 268f45cc-cd09-0410-ab3c-d52691b4dbfc