[New Multicolumn] REGRESSION: Column set styles don't update when zooming etc. https://bugs.webkit.org/show_bug.cgi?id=131777
Reviewed by Enrica Casucci.


Source/WebCore: 
Added fast/multicol/zoom-test.html

* rendering/RenderBlockFlow.cpp:
(WebCore::RenderBlockFlow::styleDidChange):
Put back in the code that propagates new styles to the flow thread and column sets
when styles change. This propagation was removed when the column span code landed.
This patch puts it back in but qualifies the types of objects that get styles
propagated to avoid accidentally propagating the wrong style to column spans.

LayoutTests: 
* fast/multicol/zoom-test-expected.html: Added.
* fast/multicol/zoom-test.html: Added.



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@167429 268f45cc-cd09-0410-ab3c-d52691b4dbfc