topContentInset does not play well with fullscreen elements https://bugs.webkit.org/show_bug.cgi?id=131955 -and corresponding- <rdar://problem/16651925>
Reviewed by Sam Weinig.


Source/WebCore: 
Calling setNeedsLayout() is not sufficient when the topContentInset has changed 
dynamically. We need to perform the layout right away and update the scrollbars. 
This works completely when the inset changes due to entering/exiting fullscreen, 
but I left a FIXME behind because it is possible to change the inset dynamically 
just through the WK2 API and end up in a situation where everything looks right 
except for the scrollbars. 
* page/FrameView.cpp:
(WebCore::FrameView::topContentInsetDidChange):
* page/FrameView.h:
* page/Page.cpp:
(WebCore::Page::setTopContentInset):

Source/WebKit2: 
If you take an element (such as a <video>) fullscreen when you have a 
topContentInset set, there will be an inset-sized gap at the top of the fullscreen 
video.

Save and restore the topContentInset whenever the scroll position is saved and 
restored.
* WebProcess/FullScreen/WebFullScreenManager.cpp:
(WebKit::WebFullScreenManager::WebFullScreenManager):
(WebKit::WebFullScreenManager::saveScrollPosition):
(WebKit::WebFullScreenManager::restoreScrollPosition):
* WebProcess/FullScreen/WebFullScreenManager.h:



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@167630 268f45cc-cd09-0410-ab3c-d52691b4dbfc