ScrollingCoordinator is unaware of topContentInset https://bugs.webkit.org/show_bug.cgi?id=132158 <rdar://problem/16706152>
Reviewed by Darin Adler.

Source/WebCore:
Test: platform/mac/fast/scrolling/scroll-select-bottom-test.html

The calculation of non-fast-scrollable regions does not currently take
the topContentOffset into account. Consequently, the logic that decides
whether to stay on the scrolling thread, or drop down to an individual
page element, can make the wrong choice. This is especially true for
small scrollable regions (such as <select> elements), where the
topContentInset may be quite close to the size of the scrollable
 element itself.

* page/scrolling/ScrollingCoordinator.cpp:
(WebCore::ScrollingCoordinator::computeNonFastScrollableRegion): Also
include the topContentInset value in our calculation.

LayoutTests:
* platform/mac/fast/scrolling/scroll-select-bottom-test-expected.txt: Added.
* platform/mac/fast/scrolling/scroll-select-bottom-test.html: Added.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@167807 268f45cc-cd09-0410-ab3c-d52691b4dbfc