[New Multicolumn] event.offsetX/offsetY don't work correctly https://bugs.webkit.org/show_bug.cgi?id=132284
Reviewed by Simon Fraser.

* rendering/RenderBoxModelObject.cpp:
(WebCore::RenderBoxModelObject::mapAbsoluteToLocalPoint):
Change the code here to only trigger for CSS Regions and not for
multi-column. Add a FIXME pointing out that the code is broken even for
CSS regions.

 * rendering/RenderMultiColumnFlowThread.cpp:
(WebCore::RenderMultiColumnFlowThread::mapAbsoluteToLocalPoint):
Override mapAbsoluteToLocalPoint in order to guess a region and attempt
to translate the point from that region to the flow thread local coordinate
space.

(WebCore::RenderMultiColumnFlowThread::physicalTranslationFromRegionToFlow):
* rendering/RenderMultiColumnFlowThread.h:
New helper function that just wraps calling the region's logical translation
function and converting to/from physical coordinates.



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@167910 268f45cc-cd09-0410-ab3c-d52691b4dbfc