[CSS Regions] Elements with overflow:auto are not painted inside regions when following a float https://bugs.webkit.org/show_bug.cgi?id=131161
Reviewed by David Hyatt.

Source/WebCore:
When a float-avoiding element (such as overflow:auto) is flowed into a flow thread after a float, its
logical left is not set on its frameRect or its layer's topLeft, but is instead kept in its RenderBoxRegionInfo
structure. In some situation, this causes paint rejection to kick-in because intersecting the layer's rect
(which DOESN'T contain the logical left from the region info object) with the computed background rect
(which contains the logical left from the region info object) results in an empty rect.

Test: fast/regions/overflow-auto-after-float.html

* rendering/RenderLayer.cpp:
(WebCore::RenderLayer::calculateClipRects):

LayoutTests:
Added test for an overflow:auto element being flowed after a float.

* fast/regions/overflow-auto-after-float-expected.html: Added.
* fast/regions/overflow-auto-after-float.html: Added.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@168288 268f45cc-cd09-0410-ab3c-d52691b4dbfc