REGRESSION (r168244): compositing/visibility/visibility-image-layers-dynamic.html failing. https://bugs.webkit.org/show_bug.cgi?id=132551
Reviewed by Tim Horton.

This just needed a re-baseline after all. The bounds have increased in size 
because the size of the layer tree dump effects the bounds! So now that we are 
printing out anchor points, the bound increase by 15 pixels for every new anchor 
point print-out.

Should not longer be marked failing.
* TestExpectations:

Added comment to the test.
* compositing/visibility/visibility-image-layers-dynamic.html:

Re-baseline.
* platform/mac/compositing/visibility/visibility-image-layers-dynamic-expected.txt:



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@168310 268f45cc-cd09-0410-ab3c-d52691b4dbfc