ChildNodesLazySnapshot::m_childNodes needs two heap allocations. <https://webkit.org/b/132616>
This seems a bit silly, since ChildNodesLazySnapshot is always
allocated on the stack.

Change it to be a bool + a Vector instead of using the nullity
as a presence indicator. This way we also don't end up with an
out-of-line ~OwnPtr for the Vector.

Reviewed by Geoff Garen.

* dom/ContainerNode.h:
(WebCore::ChildNodesLazySnapshot::ChildNodesLazySnapshot):
(WebCore::ChildNodesLazySnapshot::nextNode):
(WebCore::ChildNodesLazySnapshot::takeSnapshot):
(WebCore::ChildNodesLazySnapshot::hasSnapshot):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@168383 268f45cc-cd09-0410-ab3c-d52691b4dbfc