Occasional crash under AsyncScrollingCoordinator::frameViewRootLayerDidChange() on history navigation https://bugs.webkit.org/show_bug.cgi?id=132723
Reviewed by Sam Weinig.

Crash reports suggest that when we call AsyncScrollingCoordinator::frameViewRootLayerDidChange()
from HistoryController::restoreScrollPositionAndViewState(), the FrameView may not have a
scrolling node ID, which suggests that either the FrameView doesn't have a RenderView,
or compositing hasn't happened yet.

So if we enter AsyncScrollingCoordinator::frameViewRootLayerDidChange()
and the FrameView has no scrollLayerID, just return. We'll fix things up
when compositing kicks in later.

* page/scrolling/AsyncScrollingCoordinator.cpp:
(WebCore::AsyncScrollingCoordinator::frameViewRootLayerDidChange):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@168514 268f45cc-cd09-0410-ab3c-d52691b4dbfc