REGRESSION (r166853): fast/preloader/document-write.html is very flaky https://bugs.webkit.org/show_bug.cgi?id=130942
Reviewed by Anders Carlsson.

Source/WebCore:
* style/StyleResolveTree.cpp:
(WebCore::Style::suspendMemoryCacheClientCalls): Use a RefPtr to the main
frame as a weak pointer to a Page that will work unless the page is destroyed.
The old code tried to do it with a RefPtr to a document, but as the FIXME
points out, that won't work if the document is disassociated with its frame.

LayoutTests:
* platform/mac/TestExpectations: Turned the test back on.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@168580 268f45cc-cd09-0410-ab3c-d52691b4dbfc