[CSS Regions] Add ASSERT to make sure using the flowThread cache does not return incorrect results https://bugs.webkit.org/show_bug.cgi?id=132906
Reviewed by Simon Fraser.

If flowThreadContainingBlock() is called on an object which is in a different
flow thread than the one currently being laid out, this method will return an incorrect
result. I added an assertion for that to make sure we catch and treat any such scenarios.
For the moment, this assertion is only validated for regions, as multicol still has some issues.

No new tests required.

* rendering/RenderObject.cpp:
(WebCore::RenderObject::locateFlowThreadContainingBlockNoCache):
(WebCore::RenderObject::locateFlowThreadContainingBlock):
* rendering/RenderObject.h:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@168971 268f45cc-cd09-0410-ab3c-d52691b4dbfc